fix: centralize session status presentation
